requirements.txt
streamlit==1.21.0
camel-ai==0.2.2
python-dotenv==1.0.0
mistralai
git+https://github.com/lightaime/camel.git
